backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.0
pymongo==4.6.0
motor==3.3.2
redis[hiredis]==5.0.1
cassandra-driver==3.28.0
elasticsearch==8.11.0
psycopg2-binary==2.9.9

# Authentication & Security
clerk-backend-sdk==0.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing & AI
openai==1.6.0
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# Message Queues & WebSockets
aiokafka==0.8.11
websockets==12.0
python-socketio==5.10.0

# Payments
stripe==7.8.0

# HTTP Client
httpx==0.25.2

# Configuration & Environment
pydantic==2.5.1
pydantic-settings==2.1.0
python-decouple==3.8

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
isort==5.13.2
mypy==1.8.0